Create a Category Row
Landmarks displays each category in a row that scrolls horizontally. Add a new view type to represent the row, then display all the landmarks for that category in the new view.

Reuse parts of the Landmark view you created in Creating and Combining Views to create familiar previews of a landmark.